requirements.txt
asgiref==3.4.1
astroid==2.9.3
autopep8==1.6.0
boto3==1.20.49
botocore==1.23.49
certifi==2021.10.8
cloudinary==1.29.0
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
jmespath==0.10.0
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.0
platformdirs==2.5.1
psycopg2==2.9.3
pycodestyle==2.8.0
pylint==2.12.2
pylint-django==2.5.2
pylint-plugin-utils==0.7
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.13.3
